# swagger.yaml
openapi: 3.0.3
info:
  title: SCIM 2.0 Service Provider API
  description: |
    A minimal, standards-compliant SCIM 2.0 Service Provider implementation for enterprise IAM integration.
    
    ## Authentication
    
    This API supports three authentication methods (all enabled by default):
    
    1. **OAuth 2.0 / JWT Bearer Token** (Recommended)
       - Use `Authorization: Bearer <token>` header
       - Get tokens from `/api/dev/tokens` endpoint
    
    2. **HTTP Basic Authentication**
       - Use `Authorization: Basic <base64(username:password)>` header
       - Default credentials: admin/admin123, testuser/testpass123
    
    3. **Mutual TLS (mTLS)**
       - Certificate-based authentication
       - Configure client certificates
    
    ## Quick Start
    
    ```bash
    # Get authentication tokens
    curl http://localhost:5000/api/dev/tokens
    
    # Use a token
    TOKEN=$(curl -s http://localhost:5000/api/dev/tokens | jq -r '.tokens.jwt.full_access.token')
    curl -H "Authorization: Bearer $TOKEN" http://localhost:5000/scim/v2/Users
    
    # Or use Basic Auth
    curl -u admin:admin123 http://localhost:5000/scim/v2/Users
    ```
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Local development server
  - url: https://your-domain.com
    description: Production server
...
security:
  - BearerAuth: []
  - BasicAuth: []
paths:
...
  /scim/v2/Schemas/{schemaId}:
    get:
      summary: Get specific schema
      description: Returns a specific SCIM schema by ID
      tags:
        - Discovery
      security: []
      parameters:
        - name: schemaId
          in: path
          required: true
          schema:
            type: string
          example: urn:ietf:params:scim:schemas:core:2.0:User
      responses:
        '200':
          description: Schema details
          content:
            application/scim+json:
              schema:
                type: object
...
  /scim/v2/Users:
    get:
      summary: List users
      description: List all users with optional filtering and pagination
      tags:
        - Users
      parameters:
        - name: filter
          in: query
          description: SCIM filter expression
          schema:
            type: string
          example: userName eq "john.doe@example.com"
        - name: startIndex
          in: query
          description: 1-based index of the first result
          schema:
            type: integer
            default: 1
        - name: count
          in: query
          description: Number of results per page
          schema:
            type: integer
            default: 100
      responses:
        '200':
          description: List of users
          content:
            application/scim+json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      summary: Create user
      description: Create a new user
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/scim+json:
            schema:
              $ref: '#/components/schemas/User'
            example:
              schemas:
                - urn:ietf:params:scim:schemas:core:2.0:User
              userName: john.doe@example.com
              name:
                givenName: John
                familyName: Doe
              emails:
                - value: john.doe@example.com
                  type: work
                  primary: true
              active: true
              department: Engineering
      responses:
        '201':
          description: User created
          content:
            application/scim+json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /scim/v2/Users/{id}:
    get:
      summary: Get user
      description: Retrieve a specific user by ID
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: 550e8400-e29b-41d4-a716-446655440000
      responses:
        '200':
          description: User details
          content:
            application/scim+json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      summary: Replace user
      description: |
        Replace an entire user resource (SCIM 2.0 RFC 7644 Section 3.5.1).
        
        PUT performs a full replacement of the user resource. All attributes must be provided in the request body.
        Any attributes not included in the request will be removed from the resource (except immutable attributes like id and meta.created).
        
        **Key differences from PATCH:**
        - PUT replaces the entire resource
        - PATCH performs partial updates using operations
        
        **Required attributes:**
        - schemas
        - userName
        
        **Immutable attributes (preserved):**
        - id
        - meta.created
        - meta.location
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: 550e8400-e29b-41d4-a716-446655440000
      requestBody:
        required: true
        content:
          application/scim+json:
            schema:
              $ref: '#/components/schemas/User'
            example:
              schemas:
                - urn:ietf:params:scim:schemas:core:2.0:User
                - urn:ietf:params:scim:schemas:extension:enterprise:2.0:User
              userName: john.doe@example.com
              name:
                givenName: John
                familyName: Doe
              emails:
                - value: john.doe@example.com
                  type: work
                  primary: true
              active: true
              department: Engineering
              urn:ietf:params:scim:schemas:extension:enterprise:2.0:User:
                manager:
                  value: manager-user-id-123
                  displayName: Jane Manager
      responses:
        '200':
          description: User replaced successfully
          content:
            application/scim+json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

    patch:
      summary: Update user
      description: Update user attributes using SCIM PATCH operations
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/scim+json:
            schema:
              $ref: '#/components/schemas/PatchOp'
            example:
              schemas:
                - urn:ietf:params:scim:api:messages:2.0:PatchOp
              Operations:
                - op: replace
                  path: active
                  value: false
      responses:
        '200':
          description: User updated
          content:
            application/scim+json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Delete user
      description: Delete a user
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: User deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /scim/v2/Groups:
    get:
      summary: List groups
      description: List all groups with optional filtering and pagination
      tags:
        - Groups
      parameters:
        - name: filter
          in: query
          description: SCIM filter expression
          schema:
            type: string
        - name: startIndex
          in: query
          schema:
            type: integer
            default: 1
        - name: count
          in: query
          schema:
            type: integer
            default: 100
      responses:
        '200':
          description: List of groups
          content:
            application/scim+json:
              schema:
                $ref: '#/components/schemas/GroupListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      summary: Create group
      description: Create a new group
      tags:
        - Groups
      requestBody:
        required: true
        content:
          application/scim+json:
            schema:
              $ref: '#/components/schemas/Group'
            example:
              schemas:
                - urn:ietf:params:scim:schemas:core:2.0:Group
              displayName: Administrators
              members:
                - value: user-id-123
                  display: John Doe
      responses:
        '201':
          description: Group created
          content:
            application/scim+json:
              schema:
                $ref: '#/components/schemas/Group'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /scim/v2/Groups/{id}:
    get:
      summary: Get group
      description: Retrieve a specific group by ID
      tags:
        - Groups
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Group details
          content:
            application/scim+json:
              schema:
                $ref: '#/components/schemas/Group'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      summary: Update group
      description: Update group attributes using SCIM PATCH operations
      tags:
        - Groups
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/scim+json:
            schema:
              $ref: '#/components/schemas/PatchOp'
      responses:
        '200':
          description: Group updated
          content:
            application/scim+json:
              schema:
                $ref: '#/components/schemas/Group'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Delete group
      description: Delete a group
      tags:
        - Groups
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Group deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/supporting-data/roles:
    get:
      summary: List roles
      description: List all available roles (read-only)
      tags:
        - Supporting Data
      parameters:
        - name: startIndex
          in: query
          schema:
            type: integer
            default: 1
        - name: count
          in: query
          schema:
            type: integer
            default: 100
      responses:
        '200':
          description: List of roles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token authentication
    BasicAuth:
      type: http
      scheme: basic
      description: HTTP Basic Authentication (username:password)
...
  responses:
...
    Unauthorized:
      description: Unauthorized - Authentication required
      content:
        application/scim+json:
          schema:
            $ref: '#/components/schemas/Error'
...